% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMobileOperator.R
\name{classifyMobileOperator}
\alias{classifyMobileOperator}
\title{Title}
\usage{
classifyMobileOperator(df, phone_var, output_var)
}
\arguments{
\item{phone_var}{Name of the variable which contains the phone number}

\item{output_var}{Name of the output variable which should describe the donor}
}
\description{
Title
}
