% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStepDSummaryTables.R
\name{createStepDSummaryTables}
\alias{createStepDSummaryTables}
\title{Title}
\usage{
createStepDSummaryTables(d)
}
\arguments{
\item{d}{Consolidate data object}

\item{lucr_api_key}{Lucr API key to convert currencies}
}
\description{
Title
}
