% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTopupSummaryTable.R
\name{createTopupSummaryTables}
\alias{createTopupSummaryTables}
\title{Title}
\usage{
createTopupSummaryTables(d)
}
\arguments{
\item{d}{Consolidate data object}

\item{lucr_api_key}{Lucr API key to convert currencies}
}
\value{

}
\description{
Title
}
