% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRapidReportUsers.R
\name{getRapidReportUsers}
\alias{getRapidReportUsers}
\title{Title}
\usage{
getRapidReportUsers(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{}
}
\description{
Title
}
