% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportRecipients.R
\name{getReportRecipients}
\alias{getReportRecipients}
\title{Title}
\usage{
getReportRecipients(
  usergroup_uid,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2_session}{}
}
\value{

}
\description{
Title
}
