% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrgunitGroupSetsMembers.R
\name{getTopupOrgunitGroups}
\alias{getTopupOrgunitGroups}
\title{Title}
\usage{
getTopupOrgunitGroups(
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2_session}{}
}
\description{
Title
}
