% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrgunitGroupSetsMembers.R
\name{getTrainedByGroup}
\alias{getTrainedByGroup}
\title{Title}
\usage{
getTrainedByGroup(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{}
}
\description{
Title
}
