% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDHIS2.R
\name{loginToDHIS2}
\alias{loginToDHIS2}
\title{Title}
\usage{
loginToDHIS2(
  config_path = NULL,
  config_path_level = "dhis",
  username = NULL,
  password = NULL,
  base_url = NULL,
  d2_session_name = "d2_default_session",
  d2_session_envir = parent.frame()
)
}
\arguments{
\item{config_path}{Path to a configuration file}

\item{config_path_level}{Config path level}

\item{username}{Username if not using a config}

\item{password}{Password if not using a config}

\item{base_url}{Base URL if not using a config}

\item{d2_session_name}{Name of the session}

\item{d2_session_envir}{Environment of the session}
}
\value{

}
\description{
Essentially a re-export
of loginToDATIM from the datimutils package
}
