% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDHISConfig.R
\name{parseDHISConf}
\alias{parseDHISConf}
\title{Title}
\usage{
parseDHISConf(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
A dataframe of key value pairs parsed
from a DHIS2 configuration file
}
\description{
Title
}
