% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareWeeklyTopupData.R
\name{prepareWeeklyTopupData}
\alias{prepareWeeklyTopupData}
\title{Title}
\usage{
prepareWeeklyTopupData(report_date, d2_session)
}
\arguments{
\item{report_date}{Date as a character in format YYYY-MM-DD}

\item{d2_session}{d2_session object}
}
\description{
Title
}
