% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendWeeklyReportByEmail.R
\name{sendWeeklyReportByEmail}
\alias{sendWeeklyReportByEmail}
\title{Title}
\usage{
sendWeeklyReportByEmail(d, d2_session)
}
\arguments{
\item{d}{Data object}

\item{d2_session}{datimutils session}
}
\description{
Title
}
